+++ corusco-processor/src/main/java/cz/auderis/corusco/processor/CoruscoAnnotationProcessor.java
@@ -1,8 +1,11 @@
 package cz.auderis.corusco.processor;
 
-import cz.auderis.corusco.annotations.form.CoruscoForm;
 import cz.auderis.corusco.annotations.SwingCompanionPackage;
+import cz.auderis.corusco.annotations.form.CoruscoForm;
+import cz.auderis.corusco.annotations.form.SwingForm;
 import cz.auderis.corusco.annotations.table.CoruscoTable;
+import cz.auderis.corusco.annotations.table.SwingTable;
+import java.util.LinkedHashSet;
 import java.util.Set;
 import javax.annotation.processing.AbstractProcessor;
 import javax.annotation.processing.Filer;
@@ -13,9 +16,11 @@ import javax.annotation.processing.SupportedAnnotationTypes;
 import javax.annotation.processing.SupportedSourceVersion;
 import javax.lang.model.SourceVersion;
 import javax.lang.model.element.Element;
+import javax.lang.model.element.PackageElement;
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.util.Elements;
 import javax.lang.model.util.Types;
+import javax.tools.Diagnostic;
 
 /**
  * Annotation processor that generates Corusco form, table, and action metadata.
@@ -29,8 +34,10 @@ import javax.lang.model.util.Types;
  */
 @SupportedAnnotationTypes({
         "cz.auderis.corusco.annotations.form.CoruscoForm",
+        "cz.auderis.corusco.annotations.form.SwingForm",
         "cz.auderis.corusco.annotations.SwingCompanionPackage",
         "cz.auderis.corusco.annotations.table.CoruscoTable",
+        "cz.auderis.corusco.annotations.table.SwingTable",
         "cz.auderis.corusco.annotations.command.UiAction"
 })
 @SupportedSourceVersion(SourceVersion.RELEASE_25)
@@ -58,18 +65,76 @@ public final class CoruscoAnnotationProcessor extends AbstractProcessor {
 
     @Override
     public boolean process(Set<? extends TypeElement> annotations, RoundEnvironment roundEnv) {
-        for (Element element : roundEnv.getElementsAnnotatedWith(CoruscoForm.class)) {
+        FormProcessor formProcessor = new FormProcessor(elements, types, filer, messager);
+        TableProcessor tableProcessor = new TableProcessor(elements, types, filer, messager);
+        GeneratedSourceWriter formSourceWriter = new GeneratedSourceWriter(elements, filer, messager);
+        TableSourceWriter tableSourceWriter = new TableSourceWriter(elements, filer, messager);
+
+        Set<Element> coruscoForms = new LinkedHashSet<>(roundEnv.getElementsAnnotatedWith(CoruscoForm.class));
+        Set<Element> swingForms = new LinkedHashSet<>(roundEnv.getElementsAnnotatedWith(SwingForm.class));
+        for (Element element : coruscoForms) {
+            if (swingForms.contains(element)) {
+                error(element, "Use either @CoruscoForm or @SwingForm, not both");
+                continue;
+            }
             if (element instanceof TypeElement typeElement) {
-                new FormProcessor(elements, types, filer, messager).process(typeElement);
+                formProcessor.process(typeElement);
             }
         }
-        for (Element element : roundEnv.getElementsAnnotatedWith(CoruscoTable.class)) {
+        for (Element element : swingForms) {
+            if (coruscoForms.contains(element)) {
+                continue;
+            }
             if (element instanceof TypeElement typeElement) {
-                new TableProcessor(elements, types, filer, messager).process(typeElement);
+                SwingForm annotation = typeElement.getAnnotation(SwingForm.class);
+                FormSpec form = formProcessor.createSpec(typeElement, annotation.id(), "@SwingForm", false);
+                if (form != null) {
+                    formSourceWriter.writeFormSources(typeElement, form);
+                    PackageElement packageElement = elements.getPackageOf(typeElement);
+                    formSourceWriter.writeFormSwingSources(
+                            typeElement,
+                            form,
+                            packageElement.getQualifiedName().toString()
+                    );
+                }
+            }
+        }
+
+        Set<Element> coruscoTables = new LinkedHashSet<>(roundEnv.getElementsAnnotatedWith(CoruscoTable.class));
+        Set<Element> swingTables = new LinkedHashSet<>(roundEnv.getElementsAnnotatedWith(SwingTable.class));
+        for (Element element : coruscoTables) {
+            if (swingTables.contains(element)) {
+                error(element, "Use either @CoruscoTable or @SwingTable, not both");
+                continue;
+            }
+            if (element instanceof TypeElement typeElement) {
+                tableProcessor.process(typeElement);
+            }
+        }
+        for (Element element : swingTables) {
+            if (coruscoTables.contains(element)) {
+                continue;
+            }
+            if (element instanceof TypeElement typeElement) {
+                SwingTable annotation = typeElement.getAnnotation(SwingTable.class);
+                TableSpec table = tableProcessor.createSpec(typeElement, annotation.id(), "@SwingTable");
+                if (table != null) {
+                    tableSourceWriter.writeTableSources(typeElement, table);
+                    PackageElement packageElement = elements.getPackageOf(typeElement);
+                    tableSourceWriter.writeTableSwingSources(
+                            typeElement,
+                            table,
+                            packageElement.getQualifiedName().toString()
+                    );
+                }
             }
         }
         new SwingCompanionPackageProcessor(elements, types, filer, messager).process(roundEnv);
         new ActionProcessor(messager, new ActionSourceWriter(elements, filer, messager)).process(roundEnv);
         return true;
+    }
+
+    private void error(Element element, String message) {
+        messager.printMessage(Diagnostic.Kind.ERROR, message, element);
     }
 }